+++ src/app/api/reasoning/Interpreter.v1.headed.tsx
@@ -1,87 +1,97 @@
-import { dispatchMediatedEvent } from "@/app/utils";
-import { program, MachineEvent, InterpreterInput, Context } from ".";
+"use client";
+
 import { StateMachine, interpret, Interpreter } from "xstate";
-import { ForwardedRef, MutableRefObject, useCallback, forwardRef, useEffect, useState } from "react";
+import { MutableRefObject, useCallback, useEffect, useState } from "react";
+
+import { useReasonDemoDispatch, ReasonDemoActionTypes } from "@/app/context/ReasoningDemoContext";
+import { programV1, MachineEvent, InterpreterInput, Context, StateConfig } from ".";
 
 // Type guard to check if target is a MutableRefObject
 const isMutableRefObject = (ref: any): ref is MutableRefObject<EventTarget | null> => {
     return 'current' in ref;
 };
 
-function useLogic(input: InterpreterInput & { target: ForwardedRef<EventTarget | null> }) {
-    const { functions, states, target, context } = input;
+function useLogic(input: InterpreterInput) {
+    const { functions, states, context } = input;
+    const [currentStates, setCurrentStates] = useState<StateConfig[]>();
     const [interpreter, setInterpreter] = useState<Interpreter<Context, any, MachineEvent>>();
+    const [event, setEvent] = useState<MachineEvent>();
+    const dispatch = useReasonDemoDispatch();
+
     useEffect(() => {
         if (states.length === 0) {
             return;
         }
 
-        if (interpreter?.machine.states) {
-            const keys = Object.keys(interpreter.machine.states);
-            const ids = states.map(state => state.id);
-            const allKeysFound = keys.every(key => ids.includes(key));
-            if (allKeysFound) {
-                // we've already generated this machine
-                return;
+        if (interpreter?.initialized && !interpreter?.getSnapshot().done) {
+            if (event) {
+                interpreter.send(event.type, { payload: event.payload });
             }
+        } else if (currentStates !== states) {
+            const result: StateMachine<Context, any, MachineEvent> = programV1(states, functions);
+            const initialContext = context || {
+                status: 0,
+                requestId: "test",
+                stack: [],
+            };
+            const machine = result.withContext(initialContext);
+
+            const instance = interpret(machine).onTransition((state) => {
+                console.log(`onTransition called: machine: ${machine.id} state: ${state.value}`);
+                dispatch({
+                    type: ReasonDemoActionTypes.SET_STATE,
+                    value: {
+                        currentState: state.value,
+                        context: state.context,
+                    }
+                });
+                if (state.done) {
+                    console.log("Final state reached, stopping the interpreter.");
+                    instance.stop(); // Stop the interpreter when the final state is reached
+                }
+            });
+
+            //@ts-ignore
+            setInterpreter(instance);
+            setCurrentStates(states);
         }
 
-        const result: StateMachine<Context, any, MachineEvent> = program(states, functions);
-        const initialContext = context || {
-            status: 0,
-            requestId: "test",
-            stack: [],
-        };
-        const machine = result.withContext(initialContext);
-
-        const instance = interpret(machine).onTransition((state) => {
-            if (isMutableRefObject(target) && target.current) {
-                console.log(`onTransition called: state: ${state.value} dispatching TRANSITION target ${target.current}`);
-                dispatchMediatedEvent(target.current, [
-                    "TRANSITION",
-                    {
-                        state: state.value,
-                        context: state.context,
-                    },
-                ]);
-            }
-        });
-        //@ts-ignore
-        setInterpreter(instance);
-    }, [context, functions, setInterpreter, states, target, interpreter]);
+    }, [context, functions, setInterpreter, states, interpreter, event, currentStates, dispatch]);
 
 
     const callback = useCallback((event: MachineEvent) => {
         if (!event || !event.type || !event.payload) {
             return;
         }
+
+        if (interpreter?.getSnapshot().done) {
+            console.warn(`Attempted to send event "${event.type}" to a stopped service. The event was not sent.`);
+            return;
+        }
+
         console.log(`calling machineExecution.send type: ${event.type} payload: ${event.payload}`);
-        //@ts-ignore
-        interpreter?.send(event.type, { payload: event.payload });
+        setEvent(event);
     }, [interpreter]);
 
     return { callback, interpreter };
 }
 // Using forwardRef to wrap your component allows you to receive a ref from a parent component
-const InterpreterRef = forwardRef(({ functions, states, children }: InterpreterInput & { children: React.ReactNode }, ref: ForwardedRef<EventTarget>) => {
-    const { callback, interpreter } = useLogic({ functions, states, target: ref });
+export default function Interpreter({ functions, states, children }: InterpreterInput & { children: React.ReactNode }) {
+    const { callback, interpreter } = useLogic({ functions, states });
+    const dispatch = useReasonDemoDispatch();
 
-    if (isMutableRefObject(ref) && ref.current && interpreter) {
-        dispatchMediatedEvent(ref.current, [
-            "TRANSITION",
-            {
-                state: 'default',
-                context: interpreter?.machine.context,
+    useEffect(() => {
+        dispatch({
+            type: ReasonDemoActionTypes.SET_STATE,
+            value: {
                 callback,
-            },
-        ]);
-        interpreter.start();
-    }
+            }
+        });
+        if (interpreter && !interpreter.initialized) {
+            interpreter.start();
+        }
+    }, [dispatch, callback, interpreter]);
+
     // @ts-ignore
-    return <div ref={ref}>{children}</div>;
-});
-
-// Set displayName for the component for better debugging
-InterpreterRef.displayName = 'Interpreter';
-
-export default InterpreterRef;
+    return <div>{children}</div>;
+};
