+++ src/app/api/reasoning/engine.v1.ts
@@ -1,90 +1,158 @@
 import { interpret } from "xstate";
-import {
-  StateConfig,
-  EvaluationInput,
-  EvaluatorResult,
-  ReasoningEngine,
-  Context,
-  MachineEvent,
-  program as machineMacro,
-  Task,
-} from ".";
-import { createThread, sendMessage } from "@/app/api/openai";
+import { chatCompletion } from "@/app/api/openai";
 import { Thread } from "openai/resources/beta/index.mjs";
-import { programmer, solver } from "./prompts";
-import { MessageContentText } from "openai/resources/beta/threads/index.mjs";
 
-async function getThread() {
-  const { thread } = await createThread(
-    "Chemli",
-    "You are a helpful AI assistant tasked with helping reverse engineer cosmetic formulas from marketing claims.",
-  );
-  return thread;
+import {
+    StateConfig,
+    EvaluationInput,
+    EvaluatorResult,
+    ReasoningEngine,
+    programV1,
+    Prompt,
+} from ".";
+
+import { extractJsonFromBackticks } from "@/app/utils";
+
+
+async function solve(query: string, solver: Prompt): Promise<string> {
+    // TODO remove the use of the threads API and go with completions
+    const { user, system } = await solver(query);
+
+    const result = await chatCompletion({
+        messages: [
+            { role: 'system', content: system },
+            { role: 'user', content: user },
+        ],
+        model: "gpt-4-0125-preview" // gpt-4-0125-preview, gpt-4
+    });
+    const value = result;
+    return value || '';
 }
 
-async function solve(query: string, thread?: Thread): Promise<string> {
-  // TODO remove the use of the threads API and go with completions
-  if (!thread) {
-    thread = await getThread();
-  }
-  const { user, system } = await solver(query);
+async function program(query: string, functionCatalog: string, programmer: Prompt): Promise<StateConfig[]> {
+    const { user, system } = await programmer(query, functionCatalog);
 
-  const result = await sendMessage(thread, user, system, process.env.CHEMLI_SOLVER_ASSISTANT);
-  const value = (result?.content[0] as MessageContentText).text.value;
-  return value;
-}
+    const result = await chatCompletion({
+        messages: [
+            { role: 'system', content: system },
+            { role: 'user', content: user },
+        ],
+        model: "gpt-4", // gpt-4-0125-preview, gpt-4
+        //response_format: { type: "json_object" } gpt-4-0125-preview
+    });
+    const value = result || '';
+    let unwrapped = extractJsonFromBackticks(value) || value;
 
-async function program(query: string, functionCatalog: string, thread?: Thread): Promise<StateConfig[]> {
-  // TODO remove the use of the threads API and go with completions
-  if (!thread) {
-    thread = await getThread();
-  }
+    // check the quality of the result
+    try {
+        JSON.parse(unwrapped)
+    } catch (e) {
+        const result = await chatCompletion({
+            messages: [
+                { role: 'system', content: system },
+                { role: 'user', content: user },
+                {
+                    role: 'user', content: `your generated solution:
+                ${unwrapped}
+                generated the following error:
+                ${(e as Error).message}
+                Ensure the JSON is valid and does not contain any trailing commas, correct quotes, etc
+                Only respond with the updated JSON! Your response will be sent to JSON.parse
+                ` },
+            ],
+            model: "gpt-4",
+            //response_format: { type: "json_object" }
+        });
+        const value = result || '';
+        unwrapped = extractJsonFromBackticks(value) || value;
+    }
+    const states: StateConfig[] = JSON.parse(unwrapped);
 
-  const { user, system } = await programmer(query, functionCatalog);
-  // TODO consider creating a different assistant for the programmer, though it probably isn't useful without fine tuning
-  const result = await sendMessage(thread, user, system, process.env.CHEMLI_SOLVER_ASSISTANT);
-  const value = (result?.content[0] as MessageContentText).text.value;
-  const unwrapped = value.match(/```json\s+([\s\S]*?)\s+```/)?.[1] || value;
-  // TODO add a completion to check that the state ID's match the function catalog. Either the state matches or it is dropped
-  return JSON.parse(unwrapped) as StateConfig[];
+    // make sure the state ID's are valid
+    const notFound = states.map((state) => {
+        if (state.type != 'parallel' &&
+            state.id !== 'success' &&
+            state.id !== 'failure' &&
+            functionCatalog.indexOf(state.id) < 0) {
+            return state;
+        }
+        return undefined;
+    })
+        .filter((item) => item !== undefined)
+        .map((item) => item?.id);
+    if (notFound.length > 0) {
+        // TODO, return a recursive call to program if max count has not been exceeded
+        throw new Error(`Unknown state ID encountered: ${notFound.join(',')}`)
+    }
+
+    return JSON.parse(unwrapped) as StateConfig[];
 }
 
 async function evaluate(input: EvaluationInput): Promise<EvaluatorResult> {
-  try {
-    const tools = new Map<string, Task>(
-      input.states.map((state) => [
-        state.id,
-        {
-          description: state.id,
-          implementation: (context: Context, event: MachineEvent) => console.log(event),
-        },
-      ]),
-    );
-    const machine = machineMacro(input.states, tools);
+    try {
+        const machine = programV1(input.states, input.tools!);
 
-    const withContext = machine.withContext({
-      status: 0,
-      requestId: "test",
-      stack: [],
-    });
+        const withContext = machine.withContext({
+            status: 0,
+            requestId: "test",
+            stack: [],
+        });
 
-    const machineExecution = interpret(withContext);
-  } catch (e) {
+        const machineExecution = interpret(withContext);
+    } catch (e) {
+        return {
+            rating: 0,
+            error: e as Error,
+        };
+    }
+    // TODO better evaluation. For now if it compiles we are good. When we have evaluator models we'll expand
     return {
-      rating: 0,
-      error: e as Error,
+        rating: 1,
     };
-  }
-  // TODO better evaluation. For now if it compiles we are good. When we have evaluator models we'll expand
-  return {
-    rating: 1,
-  };
+}
+
+async function transition(taskList: string, currentState: string, payload: string, aiTransition: Prompt): Promise<string> {
+    const { user, system } = await aiTransition(taskList, currentState, payload);
+
+    const result = await chatCompletion({
+        messages: [
+            { role: 'system', content: system },
+            { role: 'user', content: user },
+        ],
+        model: "gpt-4-0125-preview", // gpt-4-0125-preview, gpt-4
+        //response_format: { type: "json_object" } gpt-4-0125-preview
+    });
+    let value = result!;
+    console.log(`engine.v2.ts.transition result is: ${value}`);
+    // TODO improve retry mechanism
+    if (currentState.indexOf(value) < 0) {
+        const result = await chatCompletion({
+            messages: [
+                { role: 'system', content: system },
+                { role: 'user', content: user },
+                {
+                    role: 'user', content: `your generated solution:
+                ${value}
+                does not include a valid transition ID! Make sure your are picking a transition ID from the provided state's transitions array
+                Do not be chatty!
+                ` },
+            ],
+            model: "gpt-4",
+        });
+        value = result!;
+        if (currentState.indexOf(value) < 0) {
+            throw new Error(`Invalid model response: ${value}`);
+        }
+    }
+
+    return value;
 }
 
 const implementation: ReasoningEngine = {
-  solver: { solve },
-  programmer: { program },
-  evaluator: { evaluate },
+    solver: { solve },
+    programmer: { program },
+    evaluator: { evaluate },
+    logic: { transition }
 };
 
 export default implementation;
