+++ src/app/api/reasoning/index.ts
@@ -1,5 +1,3 @@
 export * from "./types";
-export { default as engine } from "./engine.v1";
-export { default as program } from "./programmer.v1";
-export { default as engineV2 } from "./engine.v2";
-export { default as programV2 } from "./programmer.v2";
+export { default as engineV1 } from "./engine.v1";
+export { default as programV1 } from "./programmer.v1";
